backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
django-ninja==1.6.2
django-q==1.3.9
financedatabase==2.3.1
gunicorn==21.2.0
pandas==2.3.3
pgvector==0.2.4
psycopg2-binary==2.9.5
redis==3.5.3
scikit-learn==1.8.0
sentence-transformers==5.3.0